--- src/engine/ui/window.hpp
@@ -1,9 +1,0 @@
-#ifndef WINDOW_HPP
-#define WINDOW_HPP
-
-class Window {
-public:
-	unsigned int Width;
-	unsigned int Height;
-};
-#endif
